Move get_history_information into UpdateHistoryMixin and use it in user.update_self.
