+++ openslides_backend/action/actions/user/update.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict
 
 from ....models.models import User
 from ....permissions.management_levels import OrganizationManagementLevel
@@ -9,7 +9,7 @@ from ...util.default_schema import DefaultSchema
 from ...util.register import register_action
 from .create_update_permissions_mixin import CreateUpdatePermissionsMixin
 from .set_present import UserSetPresentAction
-from .user_mixin import LimitOfUserMixin, UserMixin
+from .user_mixin import LimitOfUserMixin, UpdateHistoryMixin, UserMixin
 
 
 @register_action("user.update")
@@ -18,6 +18,7 @@ class UserUpdate(
     CreateUpdatePermissionsMixin,
     UpdateAction,
     LimitOfUserMixin,
+    UpdateHistoryMixin,
 ):
     """
     Action to update a user.
@@ -102,129 +103,3 @@ class UserUpdate(
             ]
             self.execute_other_action(UserSetPresentAction, action_payload)
         return instance
-
-    def get_history_information(self) -> Optional[List[str]]:
-        informations: List[str] = []
-
-        # User updated
-        for instance in self.instances:
-            instance_fields = set(instance.keys())
-
-            update_fields = set(
-                [
-                    "title",
-                    "first_name",
-                    "last_name",
-                    "email",
-                    "username",
-                    "default_structure_level",
-                    "default_number",
-                    "default_vote_weight",
-                ]
-            )
-            if instance_fields & update_fields:
-                informations.append("User updated")
-                break
-
-        # User updated in meeting x
-        all_meetings: List[Set[str]] = []
-        for instance in self.instances:
-            meeting_ids: Set[str] = set()
-            for field in ("structure_level_$", "number_$", "vote_weight_$"):
-                if field in instance:
-                    meeting_ids.update(instance[field] or set())
-            all_meetings.append(meeting_ids)
-        if any(all_meetings):
-            self.add_to_history(
-                informations,
-                all_meetings,
-                "User updated in meeting {}",
-                "User updated in multiple meetings",
-            )
-
-        # Group x added/removed
-        all_groups_added: List[Set[int]] = []
-        all_groups_removed: List[Set[int]] = []
-        for instance in self.instances:
-            if "group_$_ids" in instance:
-                group_ids_from_instance = self.get_group_ids_from_instance(instance)
-                group_ids_from_db = self.get_group_ids_from_db(instance)
-                added = group_ids_from_instance - group_ids_from_db
-                removed = group_ids_from_db - group_ids_from_instance
-                all_groups_added.append(added)
-                all_groups_removed.append(removed)
-        check_added = any(all_groups_added)
-        check_removed = any(all_groups_removed)
-        if check_added and check_removed:
-            informations.append("Groups changed")
-        elif check_added:
-            self.add_to_history(
-                informations, all_groups_added, "Group {} added", "Groups changed"
-            )
-        elif check_removed:
-            self.add_to_history(
-                informations, all_groups_removed, "Group {} removed", "Groups changed"
-            )
-
-        # OML/CML changed
-        for instance in self.instances:
-            if (
-                "organization_management_level" in instance
-                or "committee_$_management_level" in instance
-            ):
-                informations.append("OML/CML changed")
-                break
-
-        # Set (in)active
-        for instance in self.instances:
-            if "is_active" in instance:
-                informations.append("Set (in)active")
-                break
-
-        return informations
-
-    def add_to_history(
-        self,
-        informations: List[str],
-        data: List[Set[Any]],
-        single_msg: str,
-        multi_msg: str,
-    ) -> None:
-        # assert data not empty
-        for entries in data:
-            if entries:
-                entry_id = list(entries)[0]
-                break
-
-        if all([set([entry_id]) == entries for entries in data]):
-            informations.append(single_msg.format(entry_id))
-        else:
-            informations.append(multi_msg)
-
-    def get_group_ids_from_db(self, instance: Dict[str, Any]) -> Set[int]:
-        user_fqid = fqid_from_collection_and_id("user", instance["id"])
-        user1 = self.datastore.get(user_fqid, ["group_$_ids"], use_changed_models=False)
-        if not user1.get("group_$_ids"):
-            return set()
-        # You can give partial group_$_ids in the instance.
-        # so groups of meetings, which meeting is not in instance,
-        # doesn't count.
-        fields = [
-            f"group_${meeting_id}_ids"
-            for meeting_id in user1["group_$_ids"]
-            if f"group_${meeting_id}_ids" in instance
-        ]
-        group_ids: Set[int] = set()
-        user2 = self.datastore.get(user_fqid, fields, use_changed_models=False)
-        for field in fields:
-            group_ids.update(user2.get(field) or [])
-        return group_ids
-
-    def get_group_ids_from_instance(self, instance: Dict[str, Any]) -> Set[int]:
-        fields = [
-            f"group_${meeting_id}_ids" for meeting_id in (instance["group_$_ids"] or [])
-        ]
-        group_ids: Set[int] = set()
-        for field in fields:
-            group_ids.update(instance.get(field) or [])
-        return group_ids
